Add files via upload
She Finds is an application that targets educational and informative gaps for women in STEM. It provides women with the necessary information for bursaries, mentorship, and available job positions. The website also provides easy access to opportunities presented on the app by having direct links to application forms and assists in finding compatibility for mentors and mentees.

The website was coded over a 24-hour period at the Girlcode hackathon on 24 -25th of August 2023. The team consisted of four team members and we tried to come up with a solution that would help in aiding women in the workplace.  The website was  coded in HTML and CSS. Despite having not won the solution received great praise by the judging panel.